+++ src/shared/mod.rs
@@ -202,7 +202,7 @@ pub(crate) fn cpu_model_name() -> String {
     match file {
         Ok(content) => {
             let reader = BufReader::new(content);
-            for line in reader.lines().flatten() {
+            for line in reader.lines().map_while(Result::ok) {
                 if line.starts_with("model name") {
                     return line
                         .replace("model name", "")
@@ -266,12 +266,15 @@ pub(crate) fn disk_space(path: &Path) -> Result<(u64, u64), ReadoutError> {
 
         let stats: libc::statfs = unsafe { s.assume_init() };
 
-        let disk_size = stats.f_blocks * stats.f_bsize as UInt;
+        let disk_size = stats.f_blocks as UInt * stats.f_bsize as UInt;
         let free = stats.f_bavail as UInt * stats.f_bsize as UInt;
 
         let used_byte = disk_size - free;
         let disk_size_byte = disk_size;
 
+        #[cfg(target_pointer_width = "32")]
+        return Ok((used_byte.into(), disk_size_byte.into()));
+        #[cfg(target_pointer_width = "64")]
         return Ok((used_byte, disk_size_byte));
     }
 
@@ -288,7 +291,7 @@ pub(crate) fn get_meminfo_value(value: &str) -> u64 {
     match file {
         Ok(content) => {
             let reader = BufReader::new(content);
-            for line in reader.lines().flatten() {
+            for line in reader.lines().map_while(Result::ok) {
                 if line.starts_with(value) {
                     let s_mem_kb: String = line.chars().filter(|c| c.is_ascii_digit()).collect();
                     return s_mem_kb.parse::<u64>().unwrap_or(0);